Задания выполнены: 1. Задание по стримам находится в Лямбда пакете, название: 'Stream_Hometask1' и 'Stream_Hometask1Test' 2. Код по алгоритмам с дописанной реализацией по пузырьку и вставками там же, файл называется 'Example_product_lambda'

+++ src/Lambda/Product.java
@@ -4,15 +4,15 @@ import java.util.Comparator;
 
 public class Product implements Comparable <Product> {
     String productTitle;
-    float productPrice;
+    int productPrice;
 
-    Product (String productTitle, float productPrice) {
+    Product (String productTitle, int productPrice) {
         this.productPrice = productPrice;
         this.productTitle = productTitle;
     }
 
     String getProductTitle () { return productTitle; }
-    float getProductPrice () { return productPrice; }
+    int getProductPrice () { return productPrice; }
 // ?????????????????
     public int compareByPrice(Product p) {
         return productPrice > p.getProductPrice() ? 1 : (productPrice == p.getProductPrice() ? 0 : -1);
